Class For Constrained Hiereachical Clustering

Files belonging to this class hold information about the constrained agglomerative clustering and allows one to display results graphically.

Format

A file belonging to this class is a list with elements:

merge

A (n-1) by 2 matrix. Row i of file "merge" describes the merging of clusters at step i of the clustering. If an element j in the row is negative, it means that observation -j was merged at this stage. If j is positive, it means that the merge was with the cluster formed at the (earlier) stage j of the algorithm. Thus negative entries in file "merge" indicate agglomerations of singletons, and positive entries indicate agglomerations of non-singletons.

height

A set of (n-1) non-decreasing real values. The clustering height is the value of the criterion associated with the clustering method for the particular agglomeration.

order

A vector giving the permutation of the original observations suitable for plotting, in the sense that a cluster plot using this ordering and matrix merge will not have crossing branches.

labels

Labels for the clustered objects.

method

The agglomerative clustering method that has been used.

call

The call that produced the result.

dist.method

The distance that has been used to create dissimilarity matrix "d" (only returned if the dissimilarity matrix object has a "method" attribute attached to it).

links

A copy of the list of edges (if a matrix of edges was provided to the function).

coords

A copy of the coordinates (if coordinates were provided to the function).

Details

The class inherits from hclust-class and describes the tree produced by the constrained clustering procedure.

All class members except links and coords are identical to those in hclust-class. several methods designed to process these objects are expected to also work with constr.hclust-class objects.

See also