Forward selection with multivariate Y using permutation under reducel model

Performs a forward selection by permutation of residuals under reduced model. Y can be multivariate.

forward.sel(
  Y,
  X,
  K = nrow(X) - 1,
  R2thresh = 0.99,
  adjR2thresh = 0.99,
  nperm = 999,
  R2more = 0.001,
  alpha = 0.05,
  Xscale = TRUE,
  Ycenter = TRUE,
  Yscale = FALSE,
  verbose = TRUE
)

Arguments

Y

Response data matrix with n rows and m columns containing quantitative variables

X

Explanatory data matrix with n rows and p columns containing quantitative variables

K

Maximum number of variables to be selected. The default is one minus the number of rows

R2thresh

Stop the forward selection procedure if the R-square of the model exceeds the stated value. This parameter can vary from 0.001 to 1

adjR2thresh

Stop the forward selection procedure if the adjusted R-square of the model exceeds the stated value. This parameter can take any value (positive or negative) smaller than 1

nperm

The number of permutation to be used.The default setting is 999 permutation.

R2more

Stop the forward selection procedure if the difference in model R-square with the previous step is lower than R2more. The default setting is 0.001

alpha

Significance level. Stop the forward selection procedure if the p-value of a variable is higher than alpha. The default is 0.05 is TRUE

Xscale

Standardize the variables in table X to variance 1. The default setting is TRUE

Ycenter

Center the variables in table Y. The default setting is TRUE

Yscale

Standardize the variables in table Y to variance 1. The default setting is FALSE.

verbose

If 'TRUE' more diagnostics are printed. The default setting is TRUE

Value

A dataframe with:

variables

The names of the variables

order

The order of the selection of the variables

R2

The R2 of the variable selected

R2Cum

The cumulative R2 of the variables selected

AdjR2Cum

The cumulative adjusted R2 of the variables selected

F

The F statistic

pval

The P-value statistic

Details

The forward selection will stop when either K, R2tresh, adjR2tresh, alpha and R2more has its parameter reached.

Note

Not yet implemented for CCA (weighted regression) and with covariables.

References

Canoco manual p.49

Author

Stephane Dray stephane.dray@univ-lyon1.fr

Examples

x <- matrix(rnorm(30),10,3)
y <- matrix(rnorm(50),10,5)
    
forward.sel(y,x,nperm=99, alpha = 0.5)
#> Testing variable 1
#> Testing variable 2
#> Testing variable 3
#> Procedure stopped (alpha criteria): pvalue for variable 3 is 0.880000 (> 0.500000)
#>   variables order        R2     R2Cum  AdjR2Cum        F pvalue
#> 1        V2     2 0.2409352 0.2409352 0.1460521 2.539284   0.09
#> 2        V3     3 0.2160625 0.4569976 0.3018541 2.785324   0.04