Compute the maximum distance of the minimum spanning tree based on a distance matrix

It is used to select a truncation value for the dbMEM approach. It returns the minimum value that keep all samples connected.

give.thresh(matdist)

Arguments

matdist

A distance matrix (class dist or matrix)

Value

The maximum distance in the minimum spanning tree.

Author

Stéphane Dray stephane.dray@univ-lyon1.fr

Examples

xy <- matrix(rnorm(60),30,2)
dxy <- dist(xy)
th <- give.thresh(dxy)