Global and local tests

These two Monte Carlo tests are used to assess the existence of 'global' and 'local' spatial structures, corresponding respectively to positive and negative Moran's I .

global.rtest(X, listw, k = 1, nperm = 499)

Arguments

X

a data matrix, with variables in columns

listw

a list of weights of class listw. Can be obtained easily using the function chooseCN.

k

integer: the number of highest \(R^2\) summed to form the test statistics

nperm

integer: the number of randomisations to be performed.

Value

An object of class randtest.

Details

They rely on the decomposition of a data matrix X into global and local components using multiple regression on Moran's Eigenvector Maps (MEMs). They require a data matrix (X) and a list of weights derived from a connection network. X is regressed onto global MEMs (U+) in the global test and on local ones (U-) in the local test. One mean \(R^2\) is obtained for each MEM, the k highest being summed to form the test statistic.

The reference distribution of these statistics are obtained by randomly permuting the rows of X.

These tests were originally part of the adegenet package for R.

References

Jombart, T., Devillard, S., Dufour, A.-B. and Pontier, D. 2008. Revealing cryptic spatial patterns in genetic variability by a new multivariate method. Heredity, 101, 92–103. doi: 10.1038/hdy.2008.34.

Author

Thibaut Jombart t.jombart@imperial.ac.uk

Examples

# wait for a generic dataset