Function to compute polynomial of geographical coordinates

This function computes orthgonal polynomials of geographical coordinates. Polynomials functions are orthogonal and centred for the weights defined in wt (i.e., orthogonal to wt). It is the classical approach to perform trend surface analysis.

orthobasis.poly(coords, degree = 2, wt = rep(1/nrow(coords), nrow(coords)))

Arguments

coords

either a Spatial* object or a matrix with geographic coordinates

degree

the degree of the polynomial

wt

a vector of weights. It is used to orthogonalize the polynomial functions

Value

an object of class orthobasisSp , subclass orthobasis

References

Dray S., Pélissier R., Couteron P., Fortin M.J., Legendre P., Peres-Neto P.R., Bellier E., Bivand R., Blanchet F.G., De Caceres M., Dufour A.B., Heegaard E., Jombart T., Munoz F., Oksanen J., Thioulouse J., Wagner H.H. (2012). Community ecology in the age of multivariate multiscale spatial analysis. Ecological Monographs 82, 257--275.

See also

Author

Stéphane Dray stephane.dray@univ-lyon1.fr

Examples

if(require("ade4", quietly = TRUE)){
data(mafragh, package = "ade4")
pol2 <- orthobasis.poly(mafragh$Spatial) 

if(require("adegraphics", quietly = TRUE)){
plot(pol2, mafragh$Spatial)
}
}
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'obj' in selecting a method for function 'coordinates': object 'pol2' not found