Function to display Moran's Eigenvector Maps (MEM) and other spatial orthogonal bases

This function allow to plot or map orthogonal bases

# S3 method for class 'orthobasisSp'
plot(x, SpORcoords, pos = -1, plot = TRUE, match.ID = FALSE, ...)

Arguments

x

an object of class orthobasisSp

SpORcoords

either a Spatial* object or a matrix with geographic coordinates

pos

an integer indicating the position of the environment where the data are stored, relative to the environment where the function is called. Useful only if storeData is FALSE

plot

a logical indicating if the graphics is displayed

match.ID

a logical indicating if names of geographic entities match rownames of the orthobasisSp object

...

additional graphical parameters (see adegpar and trellis.par.get)

Value

an object of class ADEgS, generated by the s.Spatial function of the adegraphics package

See also

Author

Stéphane Dray stephane.dray@univ-lyon1.fr

Examples

if(require("ade4", quietly = TRUE) & require("spdep", quietly = TRUE)){
data(mafragh)
me <- mem(nb2listw(mafragh$nb))

if(require("adegraphics", quietly = TRUE)){
plot(me[,1:6], mafragh$xy)
plot(me[,1:6], mafragh$Spatial) 
}
}
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'obj' in selecting a method for function 'coordinates': object 'me' not found